Functions for management processing parameters in YAML files
