+++ requirements.txt
@@ -7,6 +7,7 @@ numpy>=1.15
 pandas
 pillow
 pystackreg
+pyyaml
 requests
 scikit-image
 scipy
